F# index page, a proper Italian translation
The automatic translation does a very poor job at conveying the original
content.

Over time, in an attempt to encourage F# adoption by the Italian
developer community, it is my intention to work on a proper Italian
translation of the F# documentation or, at the very least, of its
beginner sections like what-is-fsharp, tour, and the various
get-started pages, etc.